+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - [8394](https://github.com/vegaprotocol/vega/issues/8394) - Get rid of spot liquidity provision commands and data structures.
 - [8402](https://github.com/vegaprotocol/vega/issues/8402) - Avoid division by 0 in market activity tracker
 - [8347](https://github.com/vegaprotocol/vega/issues/8347) - Market state (`ELS`) to be included in checkpoint data.
+- [8303](https://github.com/vegaprotocol/vega/issues/8303) - Add support for successor markets in datanode.
 
 
 ### 🐛 Fixes